% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseñoCL.R
\name{TablaAnova}
\alias{TablaAnova}
\title{Diseño Cuadrado Latino}
\usage{
TablaAnova(respuesta, tratamiento, fila, columna, data)
}
\arguments{
\item{respuesta}{(string) nombre de la variable respuesta.}

\item{tratamiento}{(string) nombre de la variable que representan los tratamientos.}

\item{fila}{(string) nombre de la variable que representan las filas.}

\item{columna}{(string) nombre de la variable que representan las columnas.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos correspondientes al análisis de varianza.
}
\description{
Obtiene la Tabla de Analisis de Varianza (ANOVA) para un Diseño en Cuadrado Latino (DCL).
}
\examples{
\dontrun{
# Ejemplo 1
# 1 Limpiar la memoria de R
rm(list = ls ())
# 2 Cargamos el paquete DCL
library(DCL)
# 3 Ruta del archivo y nombre del mismo
# ruta <- ''...C:/ruta/data.csv''
# 4 Lectura de archivo de datos
MisDatos <- read.csv("data.csv")

# 5 Ejecutamos la funcion TablaAnova
#TablaAnova(respuesta = "respuesta", tratamiento = "tratamiento", fila = "fila", columna = "columna", data = data)
}

# Ejemplo 2
# 1 Con datos acomodados en un Dataframe
MisDatos <- data.frame(
rendimiento = c(13.7, 13.2, 12.3, 12, 12, 11.6, 11.2, 11.1, 10.5, 10.3, 10.2, 7.7, 7.5, 5.9, 5.8,12.2),
edades = c(4, 4, 2, 2, 3, 1, 3, 1, 1, 3, 4, 2, 4, 3, 1, 2),
condicion = c(3, 1, 4, 2, 1, 4, 3, 2, 1, 2, 4, 1, 2, 4, 3, 3), 
tratamiento = c("b", "a" ,"b" ,"a" ,"b" ,"a" ,"a" ,"b" ,"c" ,"c" ,"c" ,"d" ,"d" ,"d" , "d", "c"))
# 2 Cargamos el paquete DCL
library(DCL)
# 3 Ejecutamos la funcion
TablaAnova(respuesta = "rendimiento", tratamiento = "tratamiento", fila = "edades", columna = "condicion", data = MisDatos)
tabla anova
}
